[Feature]:navbarui responsible - suhyuk
