+++ src/app/globals.css
@@ -4,4 +4,5 @@
 
 * {
   scroll-behavior: smooth;
+  box-sizing: border-box;
 }
